+++ lib/games/2048/components/animated_tile.dart
