+++ lib/games/2048/components/empty_board.dart
