+++ lib/games/2048/components/score_board.dart
