+++ lib/games/2048/const/colors.dart
