+++ lib/games/2048/managers/next_direction.dart
