+++ lib/games/2048/models/board.dart
