+++ lib/games/2048/models/board_adapter.dart
