+++ lib/games/minesweeper/components/info.dart
