+++ lib/games/minesweeper/theme/colors.dart
